Rename GraphTransformationRule/GraphTransformationApplicableRule to GraphTransformationPattern/GraphTransformationRule
