+++ org.emoflon.ibex.gt/src/org/emoflon/ibex/gt/api/GraphTransformationRule.java
@@ -1,57 +1,59 @@
 package org.emoflon.ibex.gt.api;
 
+import java.util.ArrayList;
 import java.util.Collection;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.function.Consumer;
-import java.util.stream.Collectors;
+import java.util.function.Predicate;
+import java.util.function.Supplier;
 
-import org.eclipse.emf.ecore.EObject;
 import org.emoflon.ibex.common.operational.IMatch;
+import org.emoflon.ibex.common.operational.PushoutApproach;
 import org.emoflon.ibex.gt.engine.GraphTransformationInterpreter;
 
 /**
- * This is the abstraction for all rules.
+ * This is the abstraction for all applicable rules. Via <code>apply()</code>
+ * and its variants, rules can be applied.
  * 
- * Concrete Implementations must have a constructor and methods which set the
- * parameters required for finding matches. In addition methods for to bind
- * context and deleted nodes to a specific value must be provided.
+ * If no pushout approach is set for the rule before calling an
+ * <code>apply</code> method, the pushout approach of the API which created the
+ * rule will be used.
+ * 
+ * Concrete Implementations of this class should provide typed methods for
+ * binding the parameters.
  * 
  * @param <M>
  *            the type of matches returned by this rule
  * @param <R>
  *            the own type
  */
-public abstract class GraphTransformationRule<M extends GraphTransformationMatch<M, R>, R extends GraphTransformationRule<M, R>> {
+public abstract class GraphTransformationRule<M extends GraphTransformationMatch<M, R>, R extends GraphTransformationPattern<M, R>>
+		extends GraphTransformationPattern<M, R> {
 	/**
-	 * The API.
+	 * The consumer for the subscription for enableAutoApply().
 	 */
-	protected GraphTransformationAPI api;
+	private Optional<Consumer<M>> autoApply = Optional.empty();
 
 	/**
-	 * The interpreter.
+	 * The pushout approach for the rule.
 	 */
-	protected GraphTransformationInterpreter interpreter;
+	private Optional<PushoutApproach> pushoutApproach = Optional.empty();
 
 	/**
-	 * The rule name
+	 * The consumers to notify whenever a rule is applied.
 	 */
-	private String ruleName;
+	private Collection<Consumer<M>> ruleApplicationConsumers = new ArrayList<Consumer<M>>();
 
 	/**
-	 * The parameters.
+	 * The number of rule applications until now.
 	 */
-	protected HashMap<String, Object> parameters = new HashMap<String, Object>();
+	private int ruleApplicationCount = 0;
 
 	/**
-	 * The mapping between consumers for typed and untyped matches.
-	 */
-	private Map<Consumer<M>, Consumer<IMatch>> consumers = new HashMap<Consumer<M>, Consumer<IMatch>>();
-
-	/**
-	 * Creates a new rule.
+	 * Creates a new executable rule.
+	 * 
+	 * Per default, the pushout approach which is set on API level is used.
 	 * 
 	 * @param api
 	 *            the API the rule belongs to
@@ -60,212 +62,216 @@ public abstract class GraphTransformationRule<M extends GraphTransformationMatch
 	 * @param ruleName
 	 *            the name of the rule
 	 */
-	public GraphTransformationRule(final GraphTransformationAPI api, final GraphTransformationInterpreter interpreter,
-			final String ruleName) {
-		this.api = api;
-		this.interpreter = interpreter;
-		this.ruleName = ruleName;
+	public GraphTransformationRule(final GraphTransformationAPI api,
+			final GraphTransformationInterpreter interpreter, final String ruleName) {
+		super(api, interpreter, ruleName);
 	}
 
 	/**
-	 * Returns the parameters.
+	 * Returns the pushout approach. If the pushout approach has not been set for
+	 * the rule, the pushout approach defaults to the one set for the API.
 	 * 
-	 * @return the parameters
+	 * @return the pushout approach
 	 */
-	public final HashMap<String, Object> getParameters() {
-		return this.parameters;
+	public final PushoutApproach getPushoutApproach() {
+		return this.pushoutApproach.orElse(this.api.getDefaultPushoutApproach());
 	}
 
 	/**
-	 * Returns the names of the parameters which can be bound for this rule.
+	 * Sets the pushout approach for the rule.
 	 * 
-	 * @return the parameter names
-	 */
-	protected abstract List<String> getParameterNames();
-
-	/**
-	 * Binds the parameters of the rule if there is a parameter of the same name in
-	 * the match.
-	 * 
-	 * @param match
-	 *            the match
+	 * @param pushoutApproach
+	 *            the pushout approach
 	 */
 	@SuppressWarnings("unchecked")
-	public final R bind(final IMatch match) {
-		this.getParameterNames().forEach(parameterName -> {
-			if (match.isInMatch(parameterName)) {
-				this.parameters.put(parameterName, (EObject) match.get(parameterName));
-			}
-		});
+	public final R setPushoutApproach(final PushoutApproach pushoutApproach) {
+		Objects.requireNonNull(pushoutApproach, "Pushout approach must not be null!");
+		this.pushoutApproach = Optional.of(pushoutApproach);
 		return (R) this;
 	}
 
 	/**
-	 * Binds the parameters of the rule if there is a parameter of the same name in
-	 * the match.
-	 * 
-	 * @param match
-	 *            the match
+	 * Sets the pushout approach for the rule to double pushout.
 	 */
-	@SuppressWarnings("unchecked")
-	public final R bind(final GraphTransformationMatch<?, ?> match) {
-		this.bind(match.toIMatch());
-		return (R) this;
+	public final R setDPO() {
+		return this.setPushoutApproach(PushoutApproach.DPO);
 	}
 
 	/**
-	 * Finds an arbitrary match for the rule.
-	 * 
-	 * @return an {@link Optional} for the match
+	 * Sets the pushout approach for the rule to single pushout.
 	 */
-	public final Optional<M> findAnyMatch() {
-		return this.interpreter.findAnyMatch(this.ruleName, this.parameters) //
+	public final R setSPO() {
+		return this.setPushoutApproach(PushoutApproach.SPO);
+	}
+
+	/**
+	 * Applies the rule on an arbitrary match if a match can be found.
+	 * 
+	 * @return an {@link Optional} for the the match after rule application
+	 */
+	public final Optional<M> apply() {
+		Optional<M> match = this.findAnyMatch();
+		if (!match.isPresent()) {
+			return Optional.empty();
+		}
+		return this.apply(match.get());
+	}
+
+	/**
+	 * Applies the rule on the given match.
+	 * 
+	 * @param match
+	 *            the match
+	 * @return an {@link Optional} for the the match after rule application
+	 */
+	public final Optional<M> apply(final M match) {
+		Objects.requireNonNull(match, "The match must not be null!");
+		Optional<M> comatch = this.interpreter.apply(match.toIMatch(), this.getPushoutApproach(), this.parameters)
 				.map(m -> this.convertMatch(m));
+		comatch.ifPresent(cm -> {
+			this.ruleApplicationConsumers.forEach(action -> action.accept(cm));
+			this.ruleApplicationCount++;
+		});
+		return comatch;
 	}
 
 	/**
-	 * Finds all matches for the rule.
+	 * Applies the rule on at most <code>max</code> arbitrary matches.
 	 * 
-	 * @return the list of matches
+	 * @param max
+	 *            the maximal number of rule applications
+	 * @return the matches after rule application
 	 */
-	public final Collection<M> findMatches() {
-		return this.interpreter.findMatches(this.ruleName, this.parameters).stream() //
-				.map(m -> this.convertMatch(m)) //
-				.collect(Collectors.toList());
+	public final Collection<M> apply(final int max) {
+		return this.apply(matches -> matches.size() < max && this.hasMatches());
 	}
 
 	/**
-	 * Finds all matches for the rule.
+	 * Applies the rule as long as the given condition is fulfilled.
+	 * 
+	 * @param condition
+	 *            the condition to check, based on the matches of successful rule
+	 *            applications so far
+	 * @return the matches after rule application
+	 */
+	public final Collection<M> apply(final Predicate<Collection<M>> condition) {
+		Collection<M> matches = new ArrayList<M>();
+		while (condition.test(matches)) {
+			this.apply().ifPresent(m -> matches.add(m));
+		}
+		return matches;
+	}
+
+	/**
+	 * Applies the rule after binding its parameters to the values given by the
+	 * match.
+	 * 
+	 * @param match
+	 *            the match
+	 * @return the match after rule application
+	 */
+	public final Optional<M> bindAndApply(final IMatch match) {
+		this.bind(match);
+		return this.apply();
+	}
+
+	/**
+	 * Applies the rule after binding its parameters to the values given by the
+	 * match.
+	 * 
+	 * @param match
+	 *            the match
+	 * @return the match after rule application
+	 */
+	public final Optional<M> bindAndApply(final GraphTransformationMatch<?, ?> match) {
+		this.bind(match);
+		return this.apply();
+	}
+
+	/**
+	 * Applies the rule after binding its parameters to the values given by the
+	 * match the supplier returns as long as the the match is not <code>null</code>.
+	 * 
+	 * @param matchSupplier
+	 *            the supplier for a match whose bindings shall be used
+	 * @return the matches after rule application
+	 */
+	public final Collection<M> bindAndApply(final Supplier<? extends GraphTransformationMatch<?, ?>> matchSupplier) {
+		Collection<M> matches = new ArrayList<M>();
+		GraphTransformationMatch<?, ?> match = matchSupplier.get();
+		while (match != null) {
+			this.bind(match);
+			this.apply().ifPresent(m -> matches.add(m));
+			match = matchSupplier.get();
+		}
+		return matches;
+	}
+
+	/**
+	 * Applies the rule automatically whenever a match is found.
+	 * 
+	 * Note: This will not work for rules which are contain only created elements as
+	 * the rule is always applicable in this case.
+	 */
+	public final void enableAutoApply() {
+		this.autoApply.ifPresent(c -> this.unsubscribeAppearing(c));
+		this.autoApply = Optional.of(m -> this.apply(m));
+		this.subscribeAppearing(autoApply.get());
+	}
+
+	/**
+	 * Stops automatic rule application whenever a match is found.
+	 * 
+	 * @throws IllegalArgumentException
+	 *             if <code>applyWheneverApplicable</code> has not been called
+	 *             before
+	 */
+	public final void disableAutoApply() {
+		if (!this.autoApply.isPresent()) {
+			throw new IllegalArgumentException("Cannot stop applyWheneverApplicable before start.");
+		}
+		this.unsubscribeAppearing(autoApply.get());
+		this.autoApply = Optional.empty();
+	}
+
+	/**
+	 * Returns the number of successful rule applications, i. e. apply() returned an
+	 * Optional which was not empty.
+	 * 
+	 * @return the number of successful rule applications
+	 */
+	public final int countRuleApplications() {
+		return this.ruleApplicationCount;
+	}
+
+	/**
+	 * Subscribe a notification when the given rule is applied.
 	 * 
 	 * @param action
-	 *            a Consumer for the matches found
+	 *            the {@link Consumer} to notify of rule applications
 	 */
-	public final void forEachMatch(final Consumer<M> action) {
-		this.findMatches().forEach(action);
+	public final void subscribeRuleApplications(final Consumer<M> action) {
+		this.ruleApplicationConsumers.add(action);
 	}
 
 	/**
-	 * Returns whether matches for the rule exist.
-	 * 
-	 * @return <code>true</code> if and only if there is at least one match
-	 */
-	public final boolean hasMatches() {
-		return this.findAnyMatch().isPresent();
-	}
-
-	/**
-	 * Returns the number of matches found for the rule.
-	 * 
-	 * @return the number of matches
-	 */
-	public final int countMatches() {
-		return this.findMatches().size();
-	}
-
-	/**
-	 * Subscribes notifications of all new matches found for the rule.
-	 * 
-	 * @param action
-	 *            the {@link Consumer} to notify
-	 */
-	public final void subscribeAppearing(final Consumer<M> action) {
-		this.interpreter.subscribeAppearing(this.ruleName, this.convertConsumer(action));
-	}
-
-	/**
-	 * Removes the subscription of notifications of all new matches for the given
-	 * {@link Consumer}.
+	 * Removes the subscriptions of notification of rule applications.
 	 * 
 	 * @param action
 	 *            the {@link Consumer} to remove
 	 */
-	public final void unsubscribeAppearing(final Consumer<M> action) {
-		if (this.consumers.containsKey(action)) {
-			this.interpreter.unsubscibeAppearing(this.ruleName, this.consumers.get(action));
+	public final void unsubscribeRuleApplications(final Consumer<M> action) {
+		if (this.ruleApplicationConsumers.contains(action)) {
+			this.ruleApplicationConsumers.remove(action);
 		} else {
 			throw new IllegalArgumentException("Cannot remove a consumer which was not registered before!");
 		}
 	}
 
 	/**
-	 * Subscribes notifications of all disappearing matches found for the rule.
-	 * 
-	 * @param action
-	 *            the {@link Consumer} to notify
+	 * Deletes all subscriptions of notifications of rule applications.
 	 */
-	public final void subscribeDisappearing(final Consumer<M> action) {
-		this.interpreter.subscribeDisappearing(this.ruleName, this.convertConsumer(action));
+	public final void unsubscribeRuleApplicationsAll() {
+		this.ruleApplicationConsumers.clear();
 	}
-
-	/**
-	 * Removes the subscription of notifications of all disappearing matches for the
-	 * given {@link Consumer}.
-	 * 
-	 * @param action
-	 *            the {@link Consumer} to remove
-	 */
-	public final void unsubscribeDisappearing(final Consumer<M> action) {
-		if (this.consumers.containsKey(action)) {
-			this.interpreter.unsubscibeDisappearing(this.ruleName, this.consumers.get(action));
-		} else {
-			throw new IllegalArgumentException("Cannot remove a consumer which was not registered before!");
-		}
-	}
-
-	/**
-	 * Subscribe a notification when the given match disappears.
-	 * 
-	 * @param match
-	 *            the match to observe
-	 * @param action
-	 *            the {@link Consumer} to notify
-	 */
-	public final void subscribeMatchDisappears(final M match, final Consumer<M> action) {
-		this.interpreter.subscribeMatchDisappears(match.toIMatch(), this.convertConsumer(action));
-	}
-
-	/**
-	 * Removes the subscription of a notification when the given match disappears.
-	 * 
-	 * @param match
-	 *            the match to observe
-	 * @param action
-	 *            the {@link Consumer} to remove
-	 */
-	public final void unsubscribeMatchDisappears(final M match, final Consumer<M> action) {
-		if (this.consumers.containsKey(action)) {
-			this.interpreter.unsubscribeMatchDisappears(match.toIMatch(), this.convertConsumer(action));
-		} else {
-			throw new IllegalArgumentException("Cannot remove a consumer which was not registered before!");
-		}
-	}
-
-	/**
-	 * Converts the consumer of a typed match to a consumer of an untyped
-	 * {@link IMatch} which converts the match and calls the consumer of the typed
-	 * match afterwards.
-	 * 
-	 * @param action
-	 *            the {@link Consumer} of a typed match
-	 * @return the {@link IMatch} Consumer
-	 */
-	private Consumer<IMatch> convertConsumer(final Consumer<M> action) {
-		Consumer<IMatch> consumer = new Consumer<IMatch>() {
-			@Override
-			public void accept(IMatch m) {
-				action.accept(convertMatch(m));
-			}
-		};
-		this.consumers.put(action, consumer);
-		return consumer;
-	}
-
-	/**
-	 * Convert the untyped to a typed match.
-	 * 
-	 * @param match
-	 *            the untyped match
-	 * @return the typed match
-	 */
-	protected abstract M convertMatch(final IMatch match);
 }
